Reviewed-on: https://git.btclock.dev/btclock/btclock_v3/pulls/17

+++ lib/btclock/nostrdisplay_handler.hpp
@@ -2,4 +2,4 @@
 #include <string>
 #include "utils.hpp"
 
-std::array<std::string, NUM_SCREENS> parseZapNotify(std::uint16_t amount, bool withSatsSymbol);
+std::array<std::string, NUM_SCREENS> parseZapNotify(std::uint64_t amount, bool withSatsSymbol);
